Reshuffle code around
Move task specific stuff into tasks/
