--- decanlp/tasks/almond.py
@@ -1,69 +1,0 @@
-#
-# Copyright (c) 2019, The Board of Trustees of the Leland Stanford Junior University
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-# * Redistributions of source code must retain the above copyright notice, this
-#   list of conditions and the following disclaimer.
-#
-# * Redistributions in binary form must reproduce the above copyright notice,
-#   this list of conditions and the following disclaimer in the documentation
-#   and/or other materials provided with the distribution.
-#
-# * Neither the name of the copyright holder nor the names of its
-#   contributors may be used to endorse or promote products derived from
-#   this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-from .base import BaseTask
-from .registry import register_task
-from ..util import generic_dataset
-
-
-@register_task('almond')
-class Almond(BaseTask):
-    def __init__(self, name, args):
-        super().__init__(name, args)
-
-        self._thingpedia = args.thingpedia
-        self._grammar = None
-
-    @property
-    def metrics(self):
-        return ['em', 'nem', 'nf1', 'fm', 'dm', 'bleu']
-
-    def get_splits(self, field, root, **kwargs):
-        return generic_dataset.Almond.splits(
-            fields=field, root=root, tokenize=self.tokenize, reverse_task=False, **kwargs)
-
-    def tokenize(self, sentence):
-        tokenized =  sentence.split(' ')
-
-        if self._grammar is None:
-            return tokenized
-
-    def detokenize(self, tokenized):
-        return ' '.join(tokenized)
-
-
-@register_task('reverse_almond')
-class ReverseAlmond(BaseTask):
-    @property
-    def metrics(self):
-        return ['bleu', 'em', 'nem', 'nf1']
-
-    def get_splits(self, field, root, **kwargs):
-        return generic_dataset.Almond.splits(
-            fields=field, root=root, reverse_task=True, **kwargs)
